refactor: moved render implementation to Joint subclasses
