+++ packages/forge2d/lib/src/dynamics/joints/friction_joint.dart
@@ -223,4 +223,17 @@ class FrictionJoint extends Joint {
   bool solvePositionConstraints(final SolverData data) {
     return true;
   }
+
+  @override
+  void render(DebugDraw debugDraw) {
+    super.render(debugDraw);
+
+    final xf1 = bodyA.transform;
+    final xf2 = bodyB.transform;
+    final x1 = xf1.p;
+    final x2 = xf2.p;
+    final color = Color3i.zero()..setFromRGBd(0.5, 0.8, 0.8);
+
+    debugDraw.drawSegment(x1, x2, color);
+  }
 }
