+++ packages/forge2d/lib/src/dynamics/joints/prismatic_joint.dart
@@ -763,4 +763,21 @@ class PrismaticJoint extends Joint {
     return linearError <= settings.linearSlop &&
         angularError <= settings.angularSlop;
   }
+
+  @override
+  void render(DebugDraw debugDraw) {
+    super.render(debugDraw);
+
+    final xf1 = bodyA.transform;
+    final xf2 = bodyB.transform;
+    final x1 = xf1.p;
+    final x2 = xf2.p;
+    final p1 = anchorA;
+    final p2 = anchorB;
+    final color = Color3i.zero()..setFromRGBd(0.5, 0.8, 0.8);
+
+    debugDraw.drawSegment(x1, p1, color);
+    debugDraw.drawSegment(p1, p2, color);
+    debugDraw.drawSegment(x2, p2, color);
+  }
 }
