+++ packages/forge2d/test/dynamics/joints/friction_joint_test.dart
@@ -1,5 +1,8 @@
 import 'package:forge2d/forge2d.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
+
+import '../../helpers/helpers.dart';
 
 void main() {
   group('FrictionJoint', () {
@@ -10,6 +13,29 @@ void main() {
         ..bodyB = Body(BodyDef(), world);
 
       expect(FrictionJoint(jointDef), isA<FrictionJoint>());
+    });
+
+    group('render', () {
+      late World world;
+      late DebugDraw debugDraw;
+
+      setUp(() {
+        world = World();
+        debugDraw = MockDebugDraw();
+
+        registerFallbackValue(Vector2.zero());
+        registerFallbackValue(Color3i.black);
+      });
+
+      test('draws a single segment', () {
+        final joint = FrictionJoint(
+          FrictionJointDef()
+            ..bodyA = Body(BodyDef(), world)
+            ..bodyB = Body(BodyDef(), world),
+        );
+        joint.render(debugDraw);
+        verify(() => debugDraw.drawSegment(any(), any(), any())).called(1);
+      });
     });
   });
 }
