+++ packages/forge2d/test/dynamics/joints/prismatic_joint_test.dart
@@ -1,5 +1,8 @@
 import 'package:forge2d/forge2d.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
+
+import '../../helpers/helpers.dart';
 
 void main() {
   group('PrismaticJoint', () {
@@ -10,6 +13,29 @@ void main() {
         ..bodyB = Body(BodyDef(), world);
 
       expect(PrismaticJoint(jointDef), isA<PrismaticJoint>());
+    });
+
+    group('render', () {
+      late World world;
+      late DebugDraw debugDraw;
+
+      setUp(() {
+        world = World();
+        debugDraw = MockDebugDraw();
+
+        registerFallbackValue(Vector2.zero());
+        registerFallbackValue(Color3i.black);
+      });
+
+      test('draws three segments', () {
+        final joint = PrismaticJoint(
+          PrismaticJointDef()
+            ..bodyA = Body(BodyDef(), world)
+            ..bodyB = Body(BodyDef(), world),
+        );
+        joint.render(debugDraw);
+        verify(() => debugDraw.drawSegment(any(), any(), any())).called(3);
+      });
     });
   });
 }
