+++ packages/forge2d/test/helpers/mocks.dart
@@ -1,0 +1,4 @@
+import 'package:forge2d/forge2d.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockDebugDraw extends Mock implements DebugDraw {}
